Added tnds-parser4.py showing route endpoints only
Some lines (e.g. A|The busway) are so complicated that the full route
diagram is hard to read. This version of the parser just shows
start and end stops for each route section.

Not that it doesn't identify start/end stops that are intermediate
stops on other route sections.